+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Spec.hs
@@ -46,6 +46,7 @@ import Cardano.Ledger.Conway.Rules (
 import Cardano.Ledger.Conway.TxCert (ConwayTxCert)
 import Cardano.Ledger.Conway.TxInfo (ConwayContextError)
 import Cardano.Ledger.Plutus (Language (..))
+import Cardano.Ledger.Plutus.Language (SLanguage (..))
 import Cardano.Ledger.Shelley.API (ApplyTx)
 import Cardano.Ledger.Shelley.LedgerState (StashedAVVMAddresses)
 import Cardano.Ledger.Shelley.Rules (
@@ -57,6 +58,7 @@ import Control.State.Transition (STS (..))
 import Data.Typeable (Typeable)
 import qualified Test.Cardano.Ledger.Alonzo.Binary.CostModelsSpec as CostModelsSpec
 import qualified Test.Cardano.Ledger.Alonzo.Binary.TxWitsSpec as TxWitsSpec
+import qualified Test.Cardano.Ledger.Babbage.TxInfoSpec as BabbageTxInfo
 import Test.Cardano.Ledger.Common
 import qualified Test.Cardano.Ledger.Conway.Binary.Regression as Regression
 import qualified Test.Cardano.Ledger.Conway.BinarySpec as Binary
@@ -72,7 +74,8 @@ import Test.Cardano.Ledger.Core.JSON (roundTripJsonEraSpec)
 
 spec ::
   forall era.
-  ( EraPlutusTxInfo PlutusV2 era
+  ( EraPlutusTxInfo PlutusV1 era
+  , EraPlutusTxInfo PlutusV2 era
   , EraPlutusTxInfo PlutusV3 era
   , RuleListEra era
   , ConwayEraImp era
@@ -128,4 +131,9 @@ spec =
     describe "TxWits" $ do
       TxWitsSpec.spec @era
     Regression.spec @era
-    TxInfo.spec @era
+    describe "TxInfo" $ do
+      TxInfo.spec @era
+      BabbageTxInfo.spec @era
+      xdescribe "PlutusV3" $ do
+        -- TODO: https://github.com/IntersectMBO/cardano-ledger/issues/5209
+        BabbageTxInfo.txInfoSpecV2 @era SPlutusV3